move some helpers to study-metadata
